adds user-centric delete co-authored by: Erika Helm erikadphelm@gmail.com, Lewis Graves Lewisg735@gmail.com

+++ controllers/api/posts.js
@@ -14,7 +14,12 @@ function createPost(req, res, next) {
 function deletePost(req, res, next) {
     Post.findById(req.params.id)
         .then(post => {
-            return post.deleteOne()
+            if (post.owner.equals(req.user._id)) {
+                return post.deleteOne()
+            } else {
+                res.sendStatus(401)
+            }
+            
         })
         .then(() => res.sendStatus(204))
         .catch(next)
